+++ custom_components/solarcharger/datetime.py
@@ -6,18 +6,13 @@ import logging
 from homeassistant import config_entries, core
 from homeassistant.components.datetime import DateTimeEntity, DateTimeEntityDescription
 from homeassistant.config_entries import ConfigSubentry
-from homeassistant.helpers.entity import EntityCategory
+
+# from homeassistant.helpers.entity import EntityCategory
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
-from .config_utils import is_config_entity_used_as_local_device_entity
-from .const import (
-    DATETIME,
-    DOMAIN,
-    OPTION_GLOBAL_DEFAULTS_ID,
-    OPTION_NEXT_CHARGE_TIME_TRIGGER,
-)
+from .const import DATETIME, DOMAIN, OPTION_NEXT_CHARGE_TIME_TRIGGER
 from .coordinator import SolarChargerCoordinator
-from .entity import SolarChargerEntity
+from .entity import SolarChargerEntity, SolarChargerEntityType, is_create_entity
 
 # ----------------------------------------------------------------------------
 # ----------------------------------------------------------------------------
@@ -33,10 +28,11 @@ class SolarChargerDateTimeEntity(SolarChargerEntity, DateTimeEntity):
         self,
         config_item: str,
         subentry: ConfigSubentry,
+        entity_type: SolarChargerEntityType,
         desc: DateTimeEntityDescription,
     ) -> None:
         """Initialize the datetime."""
-        SolarChargerEntity.__init__(self, config_item, subentry)
+        SolarChargerEntity.__init__(self, config_item, subentry, entity_type)
         self.set_entity_id(DATETIME, config_item)
         self.set_entity_unique_id(DATETIME, config_item)
         self.entity_description = desc
@@ -60,19 +56,20 @@ class SolarChargerDateTimeConfigEntity(SolarChargerDateTimeEntity):
         self,
         config_item: str,
         subentry: ConfigSubentry,
+        entity_type: SolarChargerEntityType,
         desc: DateTimeEntityDescription,
         default_val: datetime = datetime.min.replace(tzinfo=timezone.utc),
     ) -> None:
         """Initialise datetime."""
-        super().__init__(config_item, subentry, desc)
+        super().__init__(config_item, subentry, entity_type, desc)
 
-        if desc.entity_category == EntityCategory.CONFIG:
-            # Disable local device entities. User needs to manually enable if required.
-            if subentry.unique_id != OPTION_GLOBAL_DEFAULTS_ID:
-                if not is_config_entity_used_as_local_device_entity(
-                    subentry, config_item
-                ):
-                    self._attr_entity_registry_enabled_default = False
+        # if desc.entity_category == EntityCategory.CONFIG:
+        #     # Disable local device entities. User needs to manually enable if required.
+        #     if subentry.unique_id != OPTION_GLOBAL_DEFAULTS_ID:
+        #         if not is_config_entity_used_as_local_device_entity(
+        #             subentry, config_item
+        #         ):
+        #             self._attr_entity_registry_enabled_default = False
 
         self._attr_has_entity_name = True
         self._attr_native_value = default_val
@@ -85,7 +82,9 @@ class SolarChargerDateTimeConfigEntity(SolarChargerDateTimeEntity):
 
 # ----------------------------------------------------------------------------
 # ----------------------------------------------------------------------------
-CONFIG_DATETIME_LIST: tuple[tuple[str, DateTimeEntityDescription], ...] = (
+CONFIG_DATETIME_LIST: tuple[
+    tuple[str, SolarChargerEntityType, DateTimeEntityDescription], ...
+] = (
     #####################################
     # Control entities
     # Must haves, ie. not hidden for all
@@ -93,6 +92,7 @@ CONFIG_DATETIME_LIST: tuple[tuple[str, DateTimeEntityDescription], ...] = (
     #####################################
     (
         OPTION_NEXT_CHARGE_TIME_TRIGGER,
+        SolarChargerEntityType.LOCAL_DEFAULT,
         DateTimeEntityDescription(
             key=OPTION_NEXT_CHARGE_TIME_TRIGGER,
         ),
@@ -119,15 +119,18 @@ async def async_setup_entry(
     coordinator: SolarChargerCoordinator = hass.data[DOMAIN][config_entry.entry_id]
 
     for subentry in config_entry.subentries.values():
+        # For both global default and charger subentries
         datetimes = {}
-        for config_item, entity_description in CONFIG_DATETIME_LIST:
-            datetimes[config_item] = SolarChargerDateTimeConfigEntity(
-                config_item, subentry, entity_description
-            )
-        coordinator.charge_controls[subentry.subentry_id].datetimes = datetimes
+        for config_item, entity_type, entity_description in CONFIG_DATETIME_LIST:
+            if is_create_entity(subentry, entity_type):
+                datetimes[config_item] = SolarChargerDateTimeConfigEntity(
+                    config_item, subentry, entity_type, entity_description
+                )
 
-        async_add_entities(
-            datetimes.values(),
-            update_before_add=False,
-            config_subentry_id=subentry.subentry_id,
-        )
+        if len(datetimes) > 0:
+            coordinator.charge_controls[subentry.subentry_id].datetimes = datetimes
+            async_add_entities(
+                datetimes.values(),
+                update_before_add=False,
+                config_subentry_id=subentry.subentry_id,
+            )